+++ examples/resources/nxos_subinterface_vrf/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_subinterface_vrf.example "sys/intf/encrtd-[eth1/10.124]/rtvrfMbr"