+++ internal/provider/provider.go
@@ -238,6 +238,7 @@ func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceT
 		"nxos_loopback_interface_vrf": resourceLoopbackInterfaceVRFType{},
 		"nxos_physical_interface_vrf": resourcePhysicalInterfaceVRFType{},
 		"nxos_subinterface":           resourceSubinterfaceType{},
+		"nxos_subinterface_vrf":       resourceSubinterfaceVRFType{},
 	}, nil
 }
 
@@ -261,6 +262,7 @@ func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSou
 		"nxos_loopback_interface_vrf": dataSourceLoopbackInterfaceVRFType{},
 		"nxos_physical_interface_vrf": dataSourcePhysicalInterfaceVRFType{},
 		"nxos_subinterface":           dataSourceSubinterfaceType{},
+		"nxos_subinterface_vrf":       dataSourceSubinterfaceVRFType{},
 	}, nil
 }