+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add nxos_physical_interface_vrf resource and data source
 - Add layer attribute to nxos_physical_interface resource and data source
 - Add nxos_subinterface resource and data source
+- Add nxos_subinterface_vrf resource and data source
 
 ## 0.2.3